Create reusable LineChart component and integrate it
Extracted the chart functionality into a standalone LineChart component for better modularity and reuse. Updated OptimizationCalculator to use the new LineChart component and improved type definitions in criteriaData.
